// types/mantenimientos/certificadoMetrologico.type.ts
export interface CertificadoMetrologicoRequest {

    numeroCertificado: number 

    responsable: number 

    mantenimiento: number 
}


export interface CertificadoMetrologicoResponse {

    idCertificado: number

    numeroCertificado: number 

    responsable: number 

    mantenimiento: number 
}